data-generator/requirements.txt
psycopg2-binary>=2.9.9
Faker>=22.0.0
python-dotenv>=1.0.0
